New Fonctionbnalities : Upload profile pictures for the users, New code quality improvements

+++ frontend/src/Controller/AdminController.php
@@ -5,40 +5,52 @@ namespace App\Controller;
 use App\Entity\Admin;
 use App\Form\AdminType;
 use App\Repository\AdminRepository;
+use App\Services\notificationServices\EmailNotificationService;
+use App\Services\userServices\AdminService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[Route('/admin')]
 class AdminController extends AbstractController
 {
-    #[Route('/', name: 'app_admin_index', methods: ['GET'])]
-    public function index(AdminRepository $adminRepository): Response
+    public function __construct(private AdminService $adminService)
     {
+    }
+
+    #[Route('/', name: 'app_admin_index', methods: ['GET'])]
+    public function index(): Response
+    {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN");
         return $this->render('admin/index.html.twig', [
-            'admins' => $adminRepository->findAll(),
+            'admins' => $this->adminService->getAllAdministrators(),
         ]);
     }
 
+    /**
+     * @throws TransportExceptionInterface
+     */
     #[Route('/new', name: 'app_admin_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EmailNotificationService  $emailNotificationService): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN");
         $admin = new Admin();
         $form = $this->createForm(AdminType::class, $admin);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->persist($admin);
-            $entityManager->flush();
-
+            $data = $this->adminService->saveAdministrator($admin);
+            $this->addFlash("success", $data);
+            $emailNotificationService->sendMessage($admin->getEmail(),"Here are you're credentials.", $data);
             return $this->redirectToRoute('app_admin_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('admin/new.html.twig', [
             'admin' => $admin,
-            'form' => $form,
+            "adminForm" => $form->createView()
         ]);
     }
 
@@ -46,7 +58,7 @@ class AdminController extends AbstractController
     public function show(Admin $admin): Response
     {
         return $this->render('admin/show.html.twig', [
-            'admin' => $admin,
+            'user' => $admin,
         ]);
     }
 
@@ -71,6 +83,7 @@ class AdminController extends AbstractController
     #[Route('/{id}', name: 'app_admin_delete', methods: ['POST'])]
     public function delete(Request $request, Admin $admin, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted("ROLE_ADMIN");
         if ($this->isCsrfTokenValid('delete'.$admin->getId(), $request->request->get('_token'))) {
             $entityManager->remove($admin);
             $entityManager->flush();
